Move action_sections to models/body
