+++ infobip_channels/whatsapp/models/body/multi_product_message.py
@@ -3,7 +3,7 @@ from typing import List, Optional
 
 from pydantic import conlist, constr, validator
 
-from infobip_channels.whatsapp.models.action_sections import (
+from infobip_channels.whatsapp.models.body.action_sections import (
     SectionBase,
     SectionTitleValidatorMixin,
 )
